wellness_backend/requirements.txt
asgiref==3.8.1
bleach==6.1.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-database-url==1.0.3
django-environ==0.11.2
django-summernote==0.8.20.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.31.0
webencodings==0.5.1
whitenoise==6.7.0
